Added custom exceptions InvalidWeaponException, InvalidArmorException and InvalidUnequipException

+++ Assignment1/ICalculateCharacterDamage.cs
@@ -8,7 +8,7 @@ namespace Assignment1
 {
     interface ICalculateCharacterDamage
     {
-        public int GetPrimaryAttribute();
+        public double GetPrimaryAttribute();
         public double CalculateCharacterDamage();
     }
 }
